#23 Improve release
* Add minor improvements to README.md and pom files
* Rename benchmark module to assessment
