+++ .gitignore
@@ -14,5 +14,5 @@
 *.jar
 
 # Util classes for creating .PIE files
-/benchmark/src/main/java/com/github/introfog/pie/benchmark/Display.java
-/benchmark/src/main/java/com/github/introfog/pie/benchmark/Main.java
+/assessment/src/main/java/com/github/introfog/pie/assessment/Display.java
+/assessment/src/main/java/com/github/introfog/pie/assessment/Main.java
