+++ assessment/src/main/java/com/github/introfog/pie/assessment/collisions/broadphase/applier/DefaultActionApplier.java
@@ -1,0 +1,37 @@
+/*
+    Copyright 2020 Dmitry Chubrick
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+        http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+ */
+package com.github.introfog.pie.assessment.collisions.broadphase.applier;
+
+import com.github.introfog.pie.core.collisions.broadphase.AbstractBroadPhase;
+import com.github.introfog.pie.core.shape.IShape;
+
+import java.util.List;
+
+public class DefaultActionApplier implements IActionApplier {
+    protected long callCounter;
+
+    public DefaultActionApplier() {
+        callCounter = 0;
+    }
+
+    public void applyAction(List<AbstractBroadPhase> methods, List<IShape> methodShapes) {
+        domesticApplyAction(methods, methodShapes);
+        callCounter++;
+    }
+
+    protected void domesticApplyAction(List<AbstractBroadPhase> methods, List<IShape> methodShapes) {
+    }
+}
