+++ assessment/src/test/resources/com/github/introfog/pie/assessment/collisions/broadphase/Line/3000x2line+diffSize_20491collision.pie
