+++ assessment/src/test/resources/com/github/introfog/pie/assessment/collisions/broadphase/Line/500x5line_8487collision.pie
