+++ assessment/src/test/resources/com/github/introfog/pie/assessment/collisions/broadphase/Line/5x500line_22443collision.pie
