+++ assessment/src/test/resources/com/github/introfog/pie/assessment/collisions/broadphase/Line/5x500line_8487collision.pie
