+++ assessment/src/test/resources/com/github/introfog/pie/assessment/collisions/broadphase/Square/100x100square+scattered+diffSize_14344collision.pie
