+++ assessment/src/test/resources/com/github/introfog/pie/assessment/collisions/broadphase/Square/50x50square_9702collision.pie
