--- benchmark/pom.xml
@@ -1,41 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>com.github.introfog.pie</groupId>
-        <artifactId>root</artifactId>
-        <version>1.1-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>benchmark</artifactId>
-    <packaging>jar</packaging>
-
-    <properties>
-        <junit.version>4.13</junit.version>
-    </properties>
-
-    <dependencies>
-        <dependency>
-            <groupId>com.github.introfog.pie</groupId>
-            <artifactId>core</artifactId>
-            <version>1.1-SNAPSHOT</version>
-            <scope>compile</scope>
-        </dependency>
-        <dependency>
-            <groupId>com.github.introfog.pie</groupId>
-            <artifactId>test</artifactId>
-            <version>1.1-SNAPSHOT</version>
-            <scope>test</scope>
-        </dependency>
-        <dependency>
-            <groupId>junit</groupId>
-            <artifactId>junit</artifactId>
-            <version>${junit.version}</version>
-            <scope>compile</scope>
-        </dependency>
-    </dependencies>
-
-</project>
