+++ core/pom.xml
@@ -13,6 +13,8 @@
     <artifactId>core</artifactId>
     <packaging>jar</packaging>
 
+    <name>PIE-Core Core</name>
+
     <properties>
         <junit.version>4.13</junit.version>
     </properties>
